Divide classes into package duke
