+++ src/main/java/duke/Duke.java
@@ -1,3 +1,8 @@
+package duke;
+
+import duke.command.ByeCommand;
+import duke.command.Command;
+
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.Scanner;
