+++ src/main/java/duke/Event.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.time.format.DateTimeFormatter;
 import java.time.LocalDateTime;
 public class Event extends Task {
