+++ src/main/java/duke/command/Command.java
@@ -1,7 +1,12 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
 import java.io.IOException;
 
 public abstract class Command {
     public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException, IOException;
 }
-
-// ... other command classes for "mark", "unmark", etc.
