+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,3 +1,7 @@
+package duke.command;
+
+import duke.*;
+
 import java.io.IOException;
 
 public class DeleteCommand extends Command {
