+++ src/main/java/duke/command/MarkCommand.java
@@ -1,9 +1,13 @@
+package duke.command;
+
+import duke.*;
+
 import java.io.IOException;
 
-public class UnmarkCommand extends Command {
+public class MarkCommand extends Command {
     private int taskNumber; // the number of the task to be marked
 
-    public UnmarkCommand(int taskNumber) {
+    public MarkCommand(int taskNumber) {
         this.taskNumber = taskNumber;
     }
 
@@ -11,9 +15,9 @@ public class UnmarkCommand extends Command {
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException, IOException {
         try {
             Task task = tasks.get(taskNumber - 1); // arrays are 0-based, so subtract 1
-            task.unmarkAsDone();
+            task.markAsDone();
             storage.saveTasks(tasks); // Save the updated tasks to file
-            ui.showTaskUnmarked(task);
+            ui.showTaskMarked(task);
         } catch (IndexOutOfBoundsException e) {
             throw new DukeException("Invalid task number.");
         }
